+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG.md
 
+# 1.1.3 - Version 3 (Unchanged)
+
+- Adding autofoucs to textarea
+
+# 1.1.2 - Version 3 (Unchanged)
+
+- Adjusting code
+
+# 1.1.1 - Version 3
+
+- Forcing reording of ribbons
+
 # 1.1.0 - Version 2
 
 - Added editing and sorting. Introduced the checking of version number to edit the sort order if the value doesn't exist.